Documents/Full-Stack-Dockerized-Blog-Post-WebApp/backend/requirements/dev.txt
# Everything the developer needs in addition to the production requirements
-r prod.txt

# Testing
pytest==3.7.4
WebTest==2.0.30
factory-boy==2.11.1
pdbpp==0.10.0
flask_cors==3.0.8

# Lint and code style
black==19.3b0
flake8==3.7.7
flake8-blind-except==0.1.1
flake8-debugger==3.1.0
flake8-docstrings==1.3.0
flake8-isort==2.5
isort==4.3.4
pep8-naming==0.7.0
